requirements.txt
absl-py>=1.4.0
matplotlib>=3.5.0
networkx>=3.0
numpy>=1.24.1
typing_extensions>=4.1.1
